+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [1.4.1 (unreleased)](https://github.com/kdeldycke/extra-platforms/compare/v1.4.0...main)
+## [1.5.0 (unreleased)](https://github.com/kdeldycke/extra-platforms/compare/v1.4.0...main)
 
 > [!IMPORTANT]
 > This version is not released yet and is under active development.